Completed a demo to draw lines on a graphical display.
Bugfixes:
* writeline function was changing the vectors passed as pointers, which meant
  the vectors were changed unexpectedly in the calling function. I changed writeline
  to make copies of the vectors that can be modified in its local stack frame.
* vector util functions (swap axes and swap vectors) needed to be converted to using
  16 bit integers; they were using 8 bit temp variables which resulted in high bits
  of x or y coordinates > 255 being unexpectedly discarded

+++ lcd-panel-intro/vectors.c
@@ -15,7 +15,7 @@
 swap_axes (v)
     vector_t *v;
 {
-    uint8_t temp = v->x;
+    uint16_t temp = v->x;
     v->x = v->y;
     v->y = temp;
 }
@@ -31,7 +31,7 @@ swap_vectors (a, b)
     vector_t *a;
     vector_t *b;
 {
-    uint8_t temp;
+    uint16_t temp;
 
     temp = a->x;
     a->x = b->x;
